refactor(span): move `CompactStr` into separate file (#5609)
Pure refactor. Split `atom.rs` into 2 files.
