+++ crates/oxc_span/src/lib.rs
@@ -3,6 +3,7 @@
 //! <https://doc.rust-lang.org/beta/nightly-rustc/rustc_span>
 
 mod atom;
+mod compact_str;
 mod source_type;
 mod span;
 
@@ -10,7 +11,8 @@ pub mod cmp;
 pub mod hash;
 
 pub use crate::{
-    atom::{Atom, CompactStr, MAX_INLINE_LEN as ATOM_MAX_INLINE_LEN},
+    atom::Atom,
+    compact_str::{CompactStr, MAX_INLINE_LEN as ATOM_MAX_INLINE_LEN},
     source_type::{
         Language, LanguageVariant, ModuleKind, SourceType, UnknownExtension, VALID_EXTENSIONS,
     },
